+++ src/preload/index.ts
@@ -54,6 +54,7 @@ export interface PluginAPI {
       name: string
       base: string[]
       orderAfter: string[]
+      orderBefore: string[]
     }[]
   >
 }